Move logic into commands fns, instead of switch
Refactor command logic from `command_handlers.go` to `commands.go`.
Logic should be roundly identical except for some `break` => `return`.
`commands.go` had to be moved up a module to avoid circular deps.

+ Some minor formatting changes to improve readability.
